@@ -11,47 +11,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.0",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/forms": "~12.2.0",
-    "@angular/core": "^12.2.16",
+    "@angular/animations": "^14.2.10",
+    "@angular/common": "^14.2.10",
+    "@angular/compiler": "^14.2.10",
+    "@angular/core": "^14.2.10",
+    "@angular/forms": "^14.2.10",
     "@angular/http": "^7.2.16",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
+    "@angular/platform-browser": "^14.2.10",
+    "@angular/platform-browser-dynamic": "^14.2.10",
+    "@angular/router": "^14.2.10",
     "@nrwl/workspace": "^14.8.2",
     "@types/jest": "^27.5.2",
+    "bootstrap": "^5.2.3",
     "core-js": "^2.6.12",
-    "igniteui-angular": "12.0",
-    "primeflex": "^3.2.1",
-    "primeicons": "^5.0.0",
-    "primeng": "^12.2.3",
+    "igniteui-angular": "12.3.44",
+    "materialize-css": "^1.0.0",
+    "primeflex": "^3.3.0",
+    "primeicons": "^6.0.1",
+    "primeng": "^14.2.2",
     "rxjs": "~6.6.0",
-    "tslib": "^2.3.0",
+    "tslib": "^2.4.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.8",
-    "@angular-eslint/builder": "12.7.0",
-    "@angular-eslint/eslint-plugin": "12.7.0",
-    "@angular-eslint/eslint-plugin-template": "12.7.0",
-    "@angular-eslint/schematics": "12.7.0",
-    "@angular-eslint/template-parser": "12.7.0",
-    "@angular/cli": "^12.2.18",
-    "@angular/compiler-cli": "~12.2.0",
-    
-    "@angular/language-service": "^7.2.16",
+    "@angular-devkit/build-angular": "^14.2.9",
+    "@angular-eslint/builder": "14.1.2",
+    "@angular-eslint/eslint-plugin": "14.1.2",
+    "@angular-eslint/eslint-plugin-template": "14.1.2",
+    "@angular-eslint/schematics": "14.1.2",
+    "@angular-eslint/template-parser": "14.1.2",
+    "@angular/cli": "^14.2.9",
+    "@angular/compiler-cli": "^14.2.10",
+    "@angular/language-service": "^14.2.10",
     "@types/jasmine": "~3.8.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.28.2",
-    "@typescript-eslint/parser": "4.28.2",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
     "codelyzer": "^0.0.28",
-    "eslint": "^7.26.0",
+    "eslint": "^8.23.0",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^29.1.2",
+    "jest": "^29.3.1",
     "jest-preset-angular": "^12.2.2",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
@@ -62,6 +63,6 @@
     "protractor": "^7.0.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~4.3.5"
+    "typescript": "~4.6.4"
   }
 }
