@@ -8,8 +8,9 @@ describe('EmployeeComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ EmployeeComponent ]
-    })
+    declarations: [EmployeeComponent],
+    teardown: { destroyAfterEach: false }
+})
     .compileComponents();
   }));
 
